// src/custom-objects/enums/custom-field-type.enum.ts
export enum CustomFieldType {
  STRING = 'STRING',
  NUMBER = 'NUMBER',
  STRING_ARRAY = 'STRING_ARRAY',
  NUMBER_ARRAY = 'NUMBER_ARRAY',
  DATE = 'DATE',
  BOOLEAN = 'BOOLEAN',
  SELECT = 'SELECT',
  USER = 'USER',
  PICKLIST = 'PICKLIST',
  CUSTOM_OBJECT = 'CUSTOM_OBJECT',
  ATTACHMENT = 'ATTACHMENT',
}